app/lifecycle: reduce the cyclomatic complexity of running hooks
